PFA working with new simulation object.
